+++ node/graphql/index.ts
@@ -73,21 +73,21 @@ export const resolvers = {
     ],
   },
   Query: {
-    getStores: async (_: any, param: any, ctx: any) => {
+    getStores: async (_: any, param: any, ctx: Context) => {
       const {
         clients: { hub, sitemap, vbase },
         vtex: { logger },
       } = ctx
 
       const APP_NAME = 'store-locator'
-      const SHCEMA_NAME = 'sitemap'
+      const SCHEMA_NAME = 'sitemap'
 
       const saveInVbase = async () => {
         try {
-          const res = await sitemap.saveIndex()
+          const res: any = await sitemap.saveIndex()
 
-          if (res.data.saveIndex) {
-            await vbase.saveJSON(APP_NAME, SHCEMA_NAME, {
+          if (res?.data?.saveIndex) {
+            await vbase.saveJSON(APP_NAME, SCHEMA_NAME, {
               alreadyHasSitemap: true,
             })
 
@@ -96,7 +96,7 @@ export const resolvers = {
 
           return false
         } catch (err) {
-          logger.log({ error: err, message: 'getStores-saveInBase-error' })
+          logger.error({ error: err, message: 'getStores-saveInBase-error' })
 
           return false
         }
@@ -105,14 +105,14 @@ export const resolvers = {
       sitemap.hasSitemap().then((has: any) => {
         if (has === false) {
           vbase
-            .getJSON(APP_NAME, SHCEMA_NAME, true)
+            .getJSON(APP_NAME, SCHEMA_NAME, true)
             .then((getResponse: any) => {
-              const { alreadyHasSitemap } = getResponse
+              const { alreadyHasSitemap = false } = getResponse ?? {}
 
               !alreadyHasSitemap && saveInVbase()
             })
             .catch((err: any) =>
-              logger.log({ error: err, message: 'getStores-getJSON-error' })
+              logger.error({ error: err, message: 'getStores-getJSON-error' })
             )
         }
       })
@@ -126,17 +126,18 @@ export const resolvers = {
       const {
         data,
         data: { paging },
-      } = result
+      } = result ?? { data: { items: [] } }
 
       const pickuppoints = data.items ? data : { items: data }
 
       if (paging?.pages > 1) {
-        let i = 2
         const results = [] as any
 
-        while (i <= paging.pages) {
+        // API will return errors starting at ?page=100
+        const limitPagesTo99 = paging.pages > 99 ? 99 : paging.pages
+
+        for (let i = 2; i <= limitPagesTo99; i++) {
           results.push(hub.getStores({ ...param, page: i }))
-          i++
         }
 
         const remainingData = await Promise.all(results)
@@ -145,6 +146,9 @@ export const resolvers = {
           pickuppoints.items.push(...newResult.data.items)
         })
       }
+
+      // include for usage statistics
+      logger.info({ message: 'getStores', items: pickuppoints?.items?.length })
 
       return {
         items: pickuppoints.items